added Dependency Injection
