+++ Web/Design a Library Management System/ModelsDTO/LibraryMemberDTO.cs
@@ -1,7 +1,7 @@
 ﻿using Library;
 namespace Design_a_Library_Management_System.Model
 {
-    public class LibraryMemberDTO
+    public class LibraryMemberDTO 
     {
         public int MemberID { get; set; }
         public string Name { get; set; }
