+++ Web/Design a Library Management System/Program.cs
@@ -7,10 +7,10 @@ var builder = WebApplication.CreateBuilder(args);
 
 builder.Configuration.AddJsonFile("appsettings.json");
 string connectionString = builder.Configuration.GetConnectionString("DefaultConnection");
- builder.Services.AddSingleton(new LibraryContext(connectionString));
+builder.Services.AddSingleton(new LibraryContext(connectionString));
 builder.Services.AddControllers();
 builder.Services.AddScoped<SearchService>();
-
+builder.Services.AddScoped<IMemberService,MemberService>();
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
